requirements.txt
google-adk==0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
Deprecated>=1.2.14
litellm>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
httpx>=0.27.0
streamlit>=1.32.0
tiktoken>=0.6.0

